python/requirements.txt
paho-mqtt==1.6.1
beautifulsoup4==4.11.1
python-dotenv==0.21.0 